25-05-23 punp2002

+++ SEMIProject-boot/src/main/java/edu/kh/semi/board/controller/FreeCommentController.java
@@ -41,7 +41,7 @@ public class FreeCommentController {
 	@PostMapping("/insert")
     public int insert(@RequestBody Comment comment, HttpSession session) {
         Member loginMember = (Member) session.getAttribute("loginMember");
-        comment.setMemberNo((long) loginMember.getMemberNo());
+        comment.setMemberNo(loginMember.getMemberNo());
         return commentService.insert(comment);
     }
